fixed bug with clifford simulator, moved custom gates to extensions
- fixed bug with local_clifford_simulator calling the qubit simulator
instead
- renamed built local_qiskit_simulator executable to qiskit_simulator
- moved custom simulator gate definitions to folder
qiskit.extensions.qiskitsimulator
- added u0 gate class definition to standard extensions
- merged `exe` and `path` args to `path` for cppsimulator run command
