+++ qiskit/extensions/qiskitsimulator/__init__.py
@@ -15,31 +15,9 @@
 # limitations under the License.
 # =============================================================================
 
-"""Standard gates."""
-from .barrier import barrier
-from .ccx import ccx
-from .cswap import cswap
-from .cx import cx
-from .cxbase import cx_base
-from .cy import cy
-from .cz import cz
-from .swap import swap
-from .h import h
-from .iden import iden
-from .s import s
-from .t import t
-from .u0 import u0
-from .u1 import u1
-from .u2 import u2
-from .u3 import u3
-from .ubase import u_base
-from .x import x
-from .y import y
-from .z import z
-from .rx import rx
-from .ry import ry
-from .rz import rz
-from .cu1 import cu1
-from .ch import ch
-from .crz import crz
-from .cu3 import cu3
+"""C++ qiskit_simulator gates."""
+from .load import load
+from .noise import noise
+from .save import save
+from .uzz import uzz
+from .wait import wait
