--- src/cpp-simulator/test/inputs/multishot.json
@@ -1,119 +1,0 @@
-{
-	"id": "tests_cx",
-  "config": {
-    "shots": 100,
-    "seed": 1,
-    "simulator": "ideal",
-    "max_threads_shot": 1
-  },
-  "circuits": [
-    {
-      "name": "std_meas",
-      "compiled_circuit": {
-        "header": {
-          "clbit_labels": [["c",20]],
-          "number_of_clbits": 20,
-          "number_of_qubits": 20,
-          "qubit_labels": [["q", 0], ["q", 1], ["q", 2], ["q", 3], ["q", 4], ["q", 5], ["q", 6], ["q", 7], ["q", 8], ["q", 9],
-                           ["q", 10], ["q", 11], ["q", 12], ["q", 12], ["q", 14], ["q", 15], ["q", 16], ["q", 17], ["q", 18], ["q", 19]]
-        },
-        "operations": [
-          {"name": "h", "qubits": [0]},
-          {"name": "cx", "qubits": [0, 1]},
-          {"name": "measure", "qubits": [0], "clbits": [0]},
-          {"name": "cx", "qubits": [1, 2]},
-          {"name": "measure", "qubits": [1], "clbits": [1]},
-          {"name": "cx", "qubits": [2, 3]},
-          {"name": "measure", "qubits": [2], "clbits": [2]},
-          {"name": "cx", "qubits": [3, 4]},
-          {"name": "measure", "qubits": [3], "clbits": [3]},
-          {"name": "cx", "qubits": [4, 5]},
-          {"name": "measure", "qubits": [4], "clbits": [4]},
-          {"name": "cx", "qubits": [5, 6]},
-          {"name": "measure", "qubits": [5], "clbits": [5]},
-          {"name": "cx", "qubits": [6, 7]},
-          {"name": "measure", "qubits": [6], "clbits": [6]},
-          {"name": "cx", "qubits": [7, 8]},
-          {"name": "measure", "qubits": [7], "clbits": [7]},
-          {"name": "cx", "qubits": [8, 9]},
-          {"name": "measure", "qubits": [8], "clbits": [8]},
-          {"name": "cx", "qubits": [9, 10]},
-          {"name": "measure", "qubits": [9], "clbits": [9]},
-          {"name": "cx", "qubits": [10, 11]},
-          {"name": "measure", "qubits": [10], "clbits": [10]},
-          {"name": "cx", "qubits": [11, 12]},
-          {"name": "measure", "qubits": [11], "clbits": [11]},
-          {"name": "cx", "qubits": [12, 13]},
-          {"name": "measure", "qubits": [12], "clbits": [12]},
-          {"name": "cx", "qubits": [13, 14]},
-          {"name": "measure", "qubits": [13], "clbits": [13]},
-          {"name": "cx", "qubits": [14, 15]},
-          {"name": "measure", "qubits": [14], "clbits": [14]},
-          {"name": "cx", "qubits": [15, 16]},
-          {"name": "measure", "qubits": [15], "clbits": [15]},
-          {"name": "cx", "qubits": [16, 17]},
-          {"name": "measure", "qubits": [16], "clbits": [16]},
-          {"name": "cx", "qubits": [17, 18]},
-          {"name": "measure", "qubits": [17], "clbits": [17]},
-          {"name": "cx", "qubits": [18, 19]},
-          {"name": "measure", "qubits": [18], "clbits": [18]},
-          {"name": "measure", "qubits": [19], "clbits": [19]}
-        ]
-      }
-    },
-    {
-    	"name": "opt_meas",
-    	"compiled_circuit": {
-        "header": {
-          "clbit_labels": [["c",20]],
-          "number_of_clbits": 20,
-          "number_of_qubits": 20,
-          "qubit_labels": [["q", 0], ["q", 1], ["q", 2], ["q", 3], ["q", 4], ["q", 5], ["q", 6], ["q", 7], ["q", 8], ["q", 9],
-                           ["q", 10], ["q", 11], ["q", 12], ["q", 12], ["q", 14], ["q", 15], ["q", 16], ["q", 17], ["q", 18], ["q", 19]]
-      	},
-        "operations": [
-          {"name": "h", "qubits": [0]},
-          {"name": "cx", "qubits": [0, 1]},
-          {"name": "cx", "qubits": [1, 2]},
-          {"name": "cx", "qubits": [2, 3]},
-          {"name": "cx", "qubits": [3, 4]},
-          {"name": "cx", "qubits": [4, 5]},
-          {"name": "cx", "qubits": [5, 6]},
-          {"name": "cx", "qubits": [6, 7]},
-          {"name": "cx", "qubits": [7, 8]},
-          {"name": "cx", "qubits": [8, 9]},
-          {"name": "cx", "qubits": [9, 10]},
-          {"name": "cx", "qubits": [10, 11]},
-          {"name": "cx", "qubits": [11, 12]},
-          {"name": "cx", "qubits": [12, 13]},
-          {"name": "cx", "qubits": [13, 14]},
-          {"name": "cx", "qubits": [14, 15]},
-          {"name": "cx", "qubits": [15, 16]},
-          {"name": "cx", "qubits": [16, 17]},
-          {"name": "cx", "qubits": [17, 18]},
-          {"name": "cx", "qubits": [18, 19]},
-          {"name": "measure", "qubits": [0], "clbits": [0]},
-          {"name": "measure", "qubits": [1], "clbits": [1]},
-          {"name": "measure", "qubits": [2], "clbits": [2]},
-          {"name": "measure", "qubits": [3], "clbits": [3]},
-          {"name": "measure", "qubits": [4], "clbits": [4]},
-          {"name": "measure", "qubits": [5], "clbits": [5]},
-          {"name": "measure", "qubits": [6], "clbits": [6]},
-          {"name": "measure", "qubits": [7], "clbits": [7]},
-          {"name": "measure", "qubits": [8], "clbits": [8]},
-          {"name": "measure", "qubits": [9], "clbits": [9]},
-          {"name": "measure", "qubits": [10], "clbits": [10]},
-          {"name": "measure", "qubits": [11], "clbits": [11]},
-          {"name": "measure", "qubits": [12], "clbits": [12]},
-          {"name": "measure", "qubits": [13], "clbits": [13]},
-          {"name": "measure", "qubits": [14], "clbits": [14]},
-          {"name": "measure", "qubits": [15], "clbits": [15]},
-          {"name": "measure", "qubits": [16], "clbits": [16]},
-          {"name": "measure", "qubits": [17], "clbits": [17]},
-          {"name": "measure", "qubits": [18], "clbits": [18]},
-          {"name": "measure", "qubits": [19], "clbits": [19]}
-      	]
-    	}
-    }
-  ]
-}
