--- src/cpp-simulator/test/inputs/multishot2.json
@@ -1,33 +1,0 @@
-{
-	"id": "tests_multishot",
-  "config": {
-    "shots": 10000,
-    "seed": 1,
-    "simulator": "ideal",
-  },
-  "circuits": [
-    {
-    	"name": "opt_meas",
-    	"compiled_circuit": {
-        "header": {
-          "clbit_labels": [["c", 10]],
-          "number_of_clbits": 10,
-          "number_of_qubits": 10,
-          "qubit_labels": []
-      	},
-        "operations": [
-          {"name": "h", "qubits": [0]},
-          {"name": "cx", "qubits": [0, 1]},
-          {"name": "cx", "qubits": [1, 2]},
-          {"name": "cx", "qubits": [2, 3]},
-          {"name": "cx", "qubits": [3, 4]},
-          {"name": "cx", "qubits": [4, 5]},
-          {"name": "cx", "qubits": [5, 6]},
-          {"name": "cx", "qubits": [6, 7]},
-          {"name": "cx", "qubits": [7, 8]},
-          {"name": "cx", "qubits": [8, 9]}
-      	]
-    	}
-    }
-  ]
-}
