--- src/cpp-simulator/test/inputs/multishot3.json
@@ -1,49 +1,0 @@
-{
-	"id": "tests_cx",
-  "config": {
-    "shots": 10,
-    "seed": 1,
-    "simulator": "ideal",
-    "max_threads_shot": 1
-  },
-  "circuits": [
-    {
-      "name": "std_meas",
-      "compiled_circuit": {
-        "header": {
-          "clbit_labels": [["c",3]],
-          "number_of_clbits": 3,
-          "number_of_qubits": 3,
-          "qubit_labels": [["q", 0], ["q", 1], ["q", 2]]
-        },
-        "operations": [
-          {"name": "h", "qubits": [0]},
-          {"name": "cx", "qubits": [0, 1]},
-          {"name": "measure", "qubits": [0], "clbits": [0]},
-          {"name": "cx", "qubits": [1, 2]},
-          {"name": "measure", "qubits": [1], "clbits": [1]},
-          {"name": "measure", "qubits": [2], "clbits": [2]}
-        ]
-      }
-    },
-    {
-    	"name": "opt_meas",
-    	"compiled_circuit": {
-        "header": {
-          "clbit_labels": [["c",3]],
-          "number_of_clbits": 3,
-          "number_of_qubits": 3,
-          "qubit_labels": [["q", 0], ["q", 1], ["q", 2]]
-      	},
-        "operations": [
-          {"name": "h", "qubits": [0]},
-          {"name": "cx", "qubits": [0, 1]},
-          {"name": "cx", "qubits": [1, 2]},
-          {"name": "measure", "qubits": [0], "clbits": [0]},
-          {"name": "measure", "qubits": [1], "clbits": [1]},
-          {"name": "measure", "qubits": [2], "clbits": [2]}
-      	]
-    	}
-    }
-  ]
-}
