+++ src/cpp-simulator/test/refs/save_command.ref
@@ -1,0 +1,24 @@
+{
+    "backend": "local_qiskit_simulator",
+    "id": "test_save_command",
+    "result": [{
+            "data": {
+                "saved_quantum_states": [{
+                        "0": [[1.0, 0.0], [0.0, 0.0], [0.0, 0.0], [0.0, 0.0], [0.0, 0.0], [0.0, 0.0], [0.0, 0.0], [0.0, 0.0]],
+                        "1": [[0.707106781186548, 0.0], [0.707106781186547, 0.0], [0.0, 0.0], [0.0, 0.0], [0.0, 0.0], [0.0, 0.0], [0.0, 0.0], [0.0, 0.0]],
+                        "2": [[0.707106781186548, 0.0], [0.0, 0.0], [0.0, 0.0], [0.707106781186547, 0.0], [0.0, 0.0], [0.0, 0.0], [0.0, 0.0], [0.0, 0.0]],
+                        "3": [[0.707106781186548, 0.0], [0.0, 0.0], [0.0, 0.0], [0.0, 0.0], [0.0, 0.0], [0.0, 0.0], [0.0, 0.0], [0.707106781186547, 0.0]]
+                    }],
+                "time_taken": 0.00022
+            },
+            "name": "save_command",
+            "seed": 1,
+            "shots": 1,
+            "status": "DONE",
+            "success": true
+        }],
+    "simulator": "qubit",
+    "status": "COMPLETED",
+    "success": true,
+    "time_taken": 0.000245
+}
